feat: Refactor movie progress logic and implement Children Mode settings
This commit updates the movie progress tracking data models, integrates Children Mode configuration fetching, and updates deep link configurations.

- Renamed `durationMinutes` to `totalDuration` and `watchDurationMinutes` to `lastWatchedPosition` in `MovieResponse`, `MovieDetailModel`, and related UI components.
- Implemented `GetProgressUseCase` and integrated it into `StreamingViewModel` to sync playback progress.
- Updated `MovieRepository` and `MovieService` to handle progress reporting using `Int` types.
- Implemented `getChildrenMode` in `SubscriptionRepository` and connected it to `ChildrenModeViewModel` to fetch blocked categories and time limits.
- Updated `CategoryBar` to visualize selected blocked categories based on state.
- Updated `AndroidManifest.xml` deep links to use the production Vercel host and HTTPS scheme.
- Refactored `HistoryViewModel` to automatically load movies for retrieved dates and updated `HistoryRepository` to handle `DateResponse`.
- Improved `FavoriteViewModel` to locally remove deleted movies from the state without reloading.
- Updated "Join Family" dialog and Children Mode screen texts to English.

+++ app/src/main/java/vn/tutorial/cinemate/data/remote/requests/film/ReportProgressRequest.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ReportProgressRequest(
     @SerializedName("lastWatchedPosition")
-    val lastWatchedPosition: Long,
+    val lastWatchedPosition: Int,
     @SerializedName("totalDuration")
-    val totalDuration: Long
+    val totalDuration: Int
 )
